emodataset-on-emonet/requirements.txt
# FIRST install pytorch
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126
# in Sep 2025. Versions installed:
#   torch==2.8.0+cu126
#   torchvision==0.23.0+cu126

opencv-python==4.12.0.88
scikit-image==0.25.2
face-alignment==1.4.1

numpy==2.1.2
scikit-learn
pandas
matplotlib
tqdm # used for displaying progress bar
torchinfo
ipykernel
